fix: add reuasable hook to use stake flow actions directly in any navigator; use directly in earn nav
