+++ .changeset/fresh-dragons-smile.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+Re-use account-specific stake flow actions directly in the Earn Navigator to avoid empty stake flow navigation that was causing empty modals to appear.
